add unique constraint to user model, improve route logic, add constraint error to asyncHandler

+++ routes.js
@@ -4,6 +4,7 @@ const express = require('express');
 const { asyncHandler } = require('./middleware/asyncHandler');
 const models = require('./models');
 const { User, Course } = models;
+const auth = require('basic-auth');
 const { authenticateUser } = require('./middleware/authUser');
 
 // Construct a router instance.
@@ -11,8 +12,16 @@ const router = express.Router();
 
 // user routes
 router.get('/users', authenticateUser, asyncHandler(async (req, res) => {
-    console.log('searching for user...');
-    const user = req.currentUser;
+    const credentials = auth(req);
+    console.log('searching for user...', credentials.name);
+    const user = await User.findOne({
+        where: {
+            emailAddress: credentials.name
+        },
+        attributes: {
+            exclude: ['password', 'createdAt', 'updatedAt']
+        },
+    });
     res.json(user);
 }));
 
@@ -33,7 +42,10 @@ router.get('/courses', asyncHandler(async (req, res) => {
                 attributes: ['firstName', 'lastName'], 
                 as: 'user'
             }
-        ]
+        ],
+        attributes: {
+            exclude: ['createdAt', 'updatedAt']
+        },
     });
     res.json(courseData);
 }));
@@ -49,12 +61,15 @@ router.get('/courses/:id', asyncHandler(async (req, res) => {
                 attributes: ['firstName', 'lastName'],
                 as: 'user'
             }
-        ]
+        ],
+        attributes: {
+            exclude: ['createdAt', 'updatedAt']
+        },
     });
     res.json(course);
 }));
 
-router.post('/courses', asyncHandler(async (req, res) => {
+router.post('/courses', authenticateUser, asyncHandler(async (req, res) => {
     console.log('Creating new course...');
     const newCourse = await Course.create(req.body);
     res.set('Location', `/courses/${newCourse.id}`)
@@ -62,7 +77,7 @@ router.post('/courses', asyncHandler(async (req, res) => {
         .end();
 }));
 
-router.put('/courses/:id', asyncHandler(async (req, res) => {
+router.put('/courses/:id', authenticateUser, asyncHandler(async (req, res) => {
     console.log('Updating course ' + req.params.id + '...');
     await Course.update( req.body,
         {
@@ -73,10 +88,18 @@ router.put('/courses/:id', asyncHandler(async (req, res) => {
     res.status(204).end();
 }));
 
-router.delete('/courses/:id', asyncHandler(async (req, res) => {
-    console.log('Deleting course ' + req.params.id + '...');
-    await Course.destroy({ where: { id: req.params.id } });
-    res.status(204).end();
+router.delete('/courses/:id', authenticateUser, asyncHandler(async (req, res) => {
+    const testCourse = await Course.findAll({where:{id: req.params.id}, limit: 1});
+    console.log(testCourse);
+    if (testCourse.length > 0) {
+        console.log('course exists because length is ', testCourse.Course);
+        console.log('Deleting course ' + req.params.id + '...');
+        await Course.destroy({ where: { id: req.params.id } });
+        res.status(204).end();
+    } else {
+        res.status(400).json({message: 'course not found'}).end()
+    };
+    
 }))
 
 module.exports = router;
